# Replace MemoryMarshal calls to direct Unsafe calls