In that places we can call System.Runtime.CompilerServices.Unsafe directly instead of calling MemoryMarshal methods.