@jkotas 2020-02-23
Why is this better? If this is a performance improvement, do you have any numbers for it?