@EgorBo 2020-02-23
> Why is this better? If this is a performance improvement, do you have any numbers for it?

I guess it eliminates an extra bounds check
```csharp
var g = new byte[16];
if (BitConverter.IsLittleEndian)
{
    MemoryMarshal.TryWrite<Guid>(g, ref this);
}
```

when `TryWrite` is inlined:

```csharp
var g = new byte[16];
Span<byte> destination = g;
if (16 > (uint)destination.Length)   // this check won't be eliminated 
    return false;                    // destination.Length is always 16
Unsafe.WriteUnaligned<Guid>(ref GetReference(destination), value);
return true;
```
It'd be nice if JIT could do it though (there are issues for this)