@vanbukin 2020-02-24
@jkotas benchmark source https://gist.github.com/vanbukin/a5f6f3d294e7b11b8782d79706f9ef75
As @EgorBo say - this PR is just eliminates an extra bounds check by using code from `MemoryMarshal` directly. I think `ToByteArray` is little bit slower cuz I create a new Span. If it necessary maybe I should cast resulting byte array instead of creating a new one.
```BenchmarkDotNet=v0.12.0, OS=macOS 10.15.3 (19D76) [Darwin 19.3.0]
Intel Core i7-8850H CPU 2.60GHz (Coffee Lake), 1 CPU, 12 logical and 6 physical cores
.NET Core SDK=3.1.102
  [Host]     : .NET Core 3.1.2 (CoreCLR 4.700.20.6602, CoreFX 4.700.20.6702), X64 RyuJIT
  Job-SGTUDR : .NET Core 5.0.0 (CoreCLR 42.42.42.42424, CoreFX 42.42.42.42424), X64 RyuJIT

Toolchain=CoreRun
```
`master`
|           Method |     Mean |     Error |    StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------- |---------:|----------:|----------:|-------:|------:|------:|----------:|
|      ToByteArray | 6.195 ns | 0.0762 ns | 0.0712 ns | 0.0085 |     - |     - |      40 B |
|    TryWriteBytes | 5.237 ns | 0.0418 ns | 0.0391 ns |      - |     - |     - |         - |
| CtorReadOnlySpan | 2.642 ns | 0.0240 ns | 0.0213 ns |      - |     - |     - |         - |

`pr`
|           Method |     Mean |     Error |    StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------- |---------:|----------:|----------:|-------:|------:|------:|----------:|
|      ToByteArray | 6.239 ns | 0.0550 ns | 0.0459 ns | 0.0085 |     - |     - |      40 B |
|    TryWriteBytes | 5.102 ns | 0.0392 ns | 0.0348 ns |      - |     - |     - |         - |
| CtorReadOnlySpan | 2.178 ns | 0.0251 ns | 0.0234 ns |      - |     - |     - |         - |