@vanbukin 2020-02-24
.ToByteArray()

**PR**
```System.Guid.ToByteArray()  
mov     ecx,10h
call    qword ptr [System_Private_CoreLib!<>f__AnonymousType0`1[System.__Canon].get_messag
lea     rdx,[rax+10h]
mov     ecx,dword ptr [rax+8]
movdqu  xmm0,xmmword ptr [rsi]
movdqu  xmmword ptr [rsp+30h],xmm0
movdqu  xmm0,xmmword ptr [rsp+30h]
movdqu  xmmword ptr [rdx],xmm0
add     rsp,40h
```
**master**
```
System.Guid.ToByteArray()  
mov     ecx,10h
call    qword ptr [System_Private_CoreLib!<>f__AnonymousType0`1[System.__Canon].get_messag
lea     rdx,[rax+10h]
mov     ecx,dword ptr [rax+8]
cmp     rcx,10h
jl      System_Private_CoreLib!System.Guid.ToByteArray()$##6000F8F+0x3b
movdqu  xmm0,xmmword ptr [rsi]
movdqu  xmmword ptr [rsp+30h],xmm0
movdqu  xmm0,xmmword ptr [rsp+30h]
movdqu  xmmword ptr [rdx],xmm0
add     rsp,40h
```