@jkotas 2020-02-24
Do you see any asm difference for `Guid(ReadOnlySpan<byte> b)` on current master? I believe that the JIT is smart enough to eliminate the bounds check there.

`ToByteArray` is slow method because of it allocates a fresh array to every invocation. Saving two extra instructions by writing more unsafe code is not a good trace-off for it. It is non-measurable improvement. If you would like to fix this one, look into tweaking the JIT to eliminate the bounds check for you.